+++ android/app/src/main/kotlin/com/kylekun/one_second_diary/MainActivity.kt
@@ -1,6 +1,6 @@
 package com.kylekun.one_second_diary
 
-import io.flutter.embedding.android.FlutterActivity
+import dev.darttools.flutter_android_volume_keydown.FlutterAndroidVolumeKeydownActivity
 
-class MainActivity: FlutterActivity() {
+class MainActivity: FlutterAndroidVolumeKeydownActivity() {
 }