+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added option to change calendar days color for colorblind users
 - Added option to disable date filter in experimental file picker
 - Added reverse filter order button in experimental file picker
+- Added option to start recording by pressing volume buttons
 - Store location metadata in video when geotagging is enabled
 - Autofocus and capitalize first letter of custom locale and subtitles text fields